midi driver

+++ trunk/cpsycle/ui/src/graphics.c
@@ -1,6 +1,8 @@
 // This source is free software ; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation ; either version 2, or (at your option) any later version.
 // copyright 2000-2019 members of the psycle project http://psycle.sourceforge.net
 
+#include "../../detail/prefix.h"
+
 #include "graphics.h"
 
 void ui_graphics_init(ui_graphics* g, HDC hdc)
@@ -21,13 +23,13 @@ void ui_graphics_dispose(ui_graphics* g)
 }
 
 
-void ui_textout(ui_graphics* self, int x, int y, const char* str, int len)
+void ui_textout(ui_graphics* self, int x, int y, const char* str, size_t len)
 {	
 	TextOut(self->hdc, x, y, str, len);
 }
 
 void ui_textoutrectangle(ui_graphics* g, int x, int y, unsigned int options,
-	ui_rectangle r, const char* text, int len)
+	ui_rectangle r, const char* text, size_t len)
 {
 	RECT rect;
